requirements.txt
aiohttp-jinja2==1.2.0
aiohttp==3.6.2
aiohttp_session==2.9.0
aiohttp-security==0.4.0
jinja2==2.11.3
pyyaml>=5.1
cryptography>=3.2
websockets==8.1
Sphinx==3.0.4
sphinx_rtd_theme==0.4.3
recommonmark==0.6.0
marshmallow==3.5.1
dirhash==0.2.0
docker==4.2.0
donut-shellcode==0.9.2
marshmallow-enum==1.5.1
ldap3==2.8.1
lxml~=4.6.2  # debrief
reportlab==3.5.64  # debrief
svglib==1.0.1  # debrief
Markdown==3.3.3  # training
dnspython==2.1.0
